refactor(myreadingmanga): merge `parser` into `lib`
