+++ src/rust/multi.myreadingmanga/res/source.json
@@ -4,7 +4,7 @@
 		"lang": "multi",
 		"name": "MyReadingManga",
 		"version": 1,
-		"url": "https://myreadingmanga.info/",
+		"url": "https://myreadingmanga.info",
 		"nsfw": 2
 	}
 }
